Binded journal entry page

+++ com.stevencodeswright/DiabetesFoodJournal/DiabetesFoodJournal/Models/Dose.cs
@@ -8,7 +8,7 @@ namespace DiabetesFoodJournal.Models
     {
         public int Id { get; set; }
 
-        public int InsulinAmount { get; set; }
+        public decimal InsulinAmount { get; set; }
         public int UpFront { get; set; }
         public int Extended { get; set; }
         public decimal TimeExtended { get; set; }
